+++ .eslintrc.yml
@@ -1,0 +1,4 @@
+env:
+  es2020 : true
+  worker: true
+rules: {}
