+++ .pre-commit-config.yaml
@@ -21,3 +21,8 @@ repos:
     hooks:
       - id: flake8
         additional_dependencies: [flake8-2020]
+
+  - repo: https://github.com/pre-commit/mirrors-eslint
+    rev: v8.53.0
+    hooks:
+      - id: eslint
